+++ compiler/lib/source_map.ml
@@ -792,3 +792,12 @@ let find_in_js_file file =
       in
       Some (of_string content)
   | _ -> None
+
+module Encoding_spec = struct
+  type t =
+    { output_file : string option  (** Source map file ([None] means generate inline. *)
+    ; source_map : Standard.t  (** Source map to extend. *)
+    ; keep_empty : bool
+          (** Don't add anything to the source map (for js_of_ocaml's "empty sourcemap" option. *)
+    }
+end